// frontend/src/openapi/compassClient/models/UserDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { DaySheetDto } from './DaySheetDto';
import {
    DaySheetDtoFromJSON,
    DaySheetDtoFromJSONTyped,
    DaySheetDtoToJSON,
} from './DaySheetDto';

/**
 * 
 * @export
 * @interface UserDto
 */
export interface UserDto {
    /**
     * 
     * @type {string}
     * @memberof UserDto
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof UserDto
     */
    givenName?: string;
    /**
     * 
     * @type {string}
     * @memberof UserDto
     */
    familyName?: string;
    /**
     * 
     * @type {string}
     * @memberof UserDto
     */
    userId?: string;
    /**
     * 
     * @type {Array<DaySheetDto>}
     * @memberof UserDto
     */
    daySheets?: Array<DaySheetDto>;
    /**
     * 
     * @type {string}
     * @memberof UserDto
     */
    role?: UserDtoRoleEnum;
    /**
     * 
     * @type {boolean}
     * @memberof UserDto
     */
    deleted?: boolean;
}


/**
 * @export
 */
export const UserDtoRoleEnum = {
    SocialWorker: 'SOCIAL_WORKER',
    Participant: 'PARTICIPANT',
    Admin: 'ADMIN',
    NoRole: 'NO_ROLE'
} as const;
export type UserDtoRoleEnum = typeof UserDtoRoleEnum[keyof typeof UserDtoRoleEnum];


/**
 * Check if a given object implements the UserDto interface.
 */
export function instanceOfUserDto(value: object): boolean {
    return true;
}

export function UserDtoFromJSON(json: any): UserDto {
    return UserDtoFromJSONTyped(json, false);
}

export function UserDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserDto {
    if (json == null) {
        return json;
    }
    return {
        
        'email': json['email'] == null ? undefined : json['email'],
        'givenName': json['given_name'] == null ? undefined : json['given_name'],
        'familyName': json['family_name'] == null ? undefined : json['family_name'],
        'userId': json['user_id'] == null ? undefined : json['user_id'],
        'daySheets': json['daySheets'] == null ? undefined : ((json['daySheets'] as Array<any>).map(DaySheetDtoFromJSON)),
        'role': json['role'] == null ? undefined : json['role'],
        'deleted': json['deleted'] == null ? undefined : json['deleted'],
    };
}

export function UserDtoToJSON(value?: UserDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'email': value['email'],
        'given_name': value['givenName'],
        'family_name': value['familyName'],
        'user_id': value['userId'],
        'daySheets': value['daySheets'] == null ? undefined : ((value['daySheets'] as Array<any>).map(DaySheetDtoToJSON)),
        'role': value['role'],
        'deleted': value['deleted'],
    };
}